+++ packages/block-editor/src/components/block-draggable/test/helpers.native.js
@@ -175,7 +175,7 @@ export const getDraggableChip = ( { getByTestId } ) => {
 	let draggableChip;
 	try {
 		draggableChip = getByTestId( 'draggable-chip' );
-	} catch ( e ) {
+	} catch {
 		// NOOP.
 	}
 	return draggableChip;