+++ packages/ui/src/text/text.tsx
@@ -3,6 +3,7 @@ import clsx from 'clsx';
 import { forwardRef } from '@wordpress/element';
 import { type TextProps } from './types';
 import styles from './style.module.css';
+import defenseStyles from '../utils/css/global-css-defense.module.css';
 
 /**
  * A text component for rendering content with predefined typographic variants.
@@ -17,7 +18,13 @@ export const Text = forwardRef< HTMLSpanElement, TextProps >( function Text(
 		defaultTagName: 'span',
 		ref,
 		props: mergeProps< 'span' >( props, {
-			className: clsx( styles[ variant ], className ),
+			className: clsx(
+				styles.text,
+				variant.startsWith( 'heading-' ) && defenseStyles.heading,
+				variant.startsWith( 'body-' ) && defenseStyles.p,
+				styles[ variant ],
+				className
+			),
 		} ),
 	} );